started refactor geoman code from map to geospatial-map
