+++ libs/safe/src/lib/components/geospatial-map/geospatial-map.component.html
@@ -1,7 +1,6 @@
 <div class="flex gap-4">
   <safe-map
     (mapEvent)="handleMapEvent($event)"
-    [useGeomanTools]="true"
     [mapSettings]="mapSettings"
     [addOrDeleteLayer]="layerToAddOrDelete$ | async"
     [updateLayerOptions]="updateLayer$ | async"
