+++ libs/safe/src/lib/components/geospatial-map/geospatial-map.component.ts
@@ -4,6 +4,7 @@ import {
   EventEmitter,
   Input,
   Output,
+  ViewChild,
 } from '@angular/core';
 import { Feature, FeatureCollection } from 'geojson';
 import { BehaviorSubject } from 'rxjs';
@@ -28,6 +29,12 @@ import { IconName } from '../icon-picker/icon-picker.const';
 import { createCustomDivIcon } from '../ui/map/utils/create-div-icon';
 import { CommonModule } from '@angular/common';
 import { MapModule } from '../ui/map/map.module';
+
+import { MapComponent } from '../ui/map';
+import { AVAILABLE_GEOMAN_LANGUAGES } from '../ui/map/const/language';
+import { getMapFeature } from '../ui/map/utils/get-map-features';
+import { TranslateService } from '@ngx-translate/core';
+import { takeUntil } from 'rxjs';
 
 // type StyleChange =
 //   typeof LayerStylingComponent.prototype.edit extends EventEmitter<infer T>
@@ -75,12 +82,91 @@ export class GeospatialMapComponent
   private timeout: ReturnType<typeof setTimeout> | null = null;
   @Output() mapChange = new EventEmitter<Feature | FeatureCollection>();
 
+  @ViewChild(MapComponent) mapComponent?: MapComponent;
+
   /**
    * Component for displaying the input map
    * of the geospatial type question.
+   *
+   * @param translate
    */
-  constructor() {
+  constructor(private translate: TranslateService) {
     super();
+  }
+
+  /** Set geoman listeners */
+  private setUpPmListeners() {
+    // By default all drawn layer types have this property to false except for markers and circleMarkers
+    // https://github.com/geoman-io/leaflet-geoman#draw-mode
+    // We will enable this one for all layer types(including Markers) in order to auto blur when one marker is set
+    this.mapComponent?.map.pm.setGlobalOptions({ continueDrawing: false });
+    // updates question value on adding new shape
+    this.mapComponent?.map.on('pm:create', (l: any) => {
+      if (l.shape === 'Marker') {
+        l.layer.setIcon(
+          createCustomDivIcon({
+            icon: 'leaflet_default',
+            color: '#3388ff',
+            opacity: 1,
+            size: 24,
+          })
+        );
+        // If we add a Marker, we will disable the control to set new markers(currently we want to add just one)
+        this.mapComponent?.map.pm.Toolbar.setButtonDisabled('drawMarker', true);
+      }
+
+      // subscribe to changes on the created layers
+      l.layer.on(
+        'pm:change',
+        this.handleMapEvent({
+          type: MapEventType.MAP_CHANGE,
+          content: getMapFeature(this.mapComponent?.map),
+        })
+      );
+
+      l.layer.on('click', (e: any) => {
+        this.handleMapEvent({
+          type: MapEventType.SELECTED_LAYER,
+          content: { layer: e.target },
+        });
+      });
+    });
+
+    // updates question value on removing shapes
+    this.mapComponent?.map.on('pm:remove', () => {
+      const containsPointMarker = (feature: any) =>
+        feature.geometry.type === 'Point';
+      const content = getMapFeature(this.mapComponent?.map);
+      // If no markers, we enable the point marker control again
+      if (!content || !containsPointMarker(content)) {
+        this.mapComponent?.map.pm.Toolbar.setButtonDisabled(
+          'drawMarker',
+          false
+        );
+      }
+      this.handleMapEvent({
+        type: MapEventType.MAP_CHANGE,
+        content,
+      });
+    });
+
+    // set language
+    const setLang = (lang: string) => {
+      if (AVAILABLE_GEOMAN_LANGUAGES.includes(lang)) {
+        this.mapComponent?.map.pm.setLang(lang);
+      } else {
+        console.warn(`Language "${lang}" not supported by geoman`);
+        this.mapComponent?.map.pm.setLang('en');
+      }
+    };
+
+    setLang(this.translate.currentLang || 'en');
+
+    this.translate.onLangChange
+      .pipe(takeUntil(this.destroy$))
+      .subscribe((event) => {
+        setLang(event.lang);
+      });
   }
 
   ngAfterViewInit(): void {
@@ -116,6 +202,7 @@ export class GeospatialMapComponent
       label: '',
       layer,
     };
+    this.setUpPmListeners();
     this.addOrDeleteLayer.next({ layerData: baseLayer, isDelete: false });
     this.setDataLayers();
   }
